[refactor] custom hook을 이용한 IssueDetail 리팩토링
- custom hook을 이용한 Issue Detail 리팩토링
  Issue Detail 에서 관리하는 상태 tabIndex, crimeIndex에 연관된 상태 변경 함수가 많음.
  따라서 custom hook을 이용하여 로직을 분리해줌
